fix(auth): run full function in both workflows to avoid conflicts
